nvim: cleans up lsp stuff. Has it return functions
Updates the lspconfig lua module to return a table with the on_attach()
and (new) setup() functions. The setup() function is what gets called to
actually configure the lsp plugin with the various LSP servers. The
on_attach function is returned so it can be used in null-ls as well