nvim: updates null-ls to put temp files used for formatting in /tmp
Updates the null ls plugin to put temporary files into the /tmp
directory for the system since it defaults to putting the file next to
the file being worked with. This was causing issues for larger VHDL
files when formatting with VSG in the editor, since it would take too
long. So if it got into a weird state and didn't delete the temp file,
then vhdl_ls would see it as a duplicate, and cause an error