fix(nvim): Adds coq_nvim for VHDL temporarily since nvim-cmp freezes
This is only temporary. I am using nvim-cmp for _everything_ besides
completions in VHDL buffers, since nvim-cmp freezes the whole editor
when unisim is loaded and snippets are enabled. However, who knows, I
may switch to coq_nvim completely if I can config it as nicely as
nvim-cmp is

See:
https://github.com/hrsh7th/nvim-cmp/issues/2028
https://github.com/VHDL-LS/rust_hdl/issues/340