nixos: adds udev rule to handle eGPU
this is a very hacky way to handle the eGPU I have. Over a couple
reboots, I noticed that the GPU enumerated as card0, so gonna hardcode
this for now. This tells udev to tag that card with something that
mutter looks at to determine which GPU is the primary. With a quick
test, this made it so that the GPU was used for factorio and for
unigine-heaven (GPU benchmarking). I wanna investigate doing a better
option with udev, probably based on the PCIE ID of the GPU