[hm]: adds a desktop entry for wezterm to use nixGL on work PC
since my work PC isn't NixOS, but just has nix and home-manager
installed, it needs NixGL to handle opening nix installed OpenGL based
GUI things. So this adds a desktop entry to override the one that comes
with the package that prepends nixGL before the exec command. Based it
on the .desktop file that is in the wezterm repo