updates tree-sitter
updates tree-sitter to the latest version by changing the branch to
main, since I didn't realize they archived master but it's still the
default. This change was a complete rewrite, so this will be fun. Adds
an autocommand to actually enable tree sitter since it doesn't do it
automatically now